+++ opentps_core/opentps/core/processing/doseCalculation/_utils.py
@@ -152,7 +152,7 @@ class MCsquareSharedLib():
                 cumContValNb += contValNb
 
             A = csc_matrix((rowVal[spotInd], (spotRowInd, np.zeros((nbVal, )))), shape=(nVoxels, 1),
-                           dtype=np.float64)
+                           dtype=np.float32)
 
             if beamletMat is None:
                 beamletMat = A
